+++ internal/multislog/multislog.go
@@ -1,4 +1,4 @@
-package main
+package multislog
 
 import (
 	"context"
@@ -6,9 +6,9 @@ import (
 	"log/slog"
 )
 
-type multiOutputHandler []slog.Handler
+type Handler []slog.Handler
 
-func (m multiOutputHandler) Enabled(ctx context.Context, level slog.Level) bool {
+func (m Handler) Enabled(ctx context.Context, level slog.Level) bool {
 	for _, handler := range m {
 		if handler.Enabled(ctx, level) {
 			return true
@@ -17,7 +17,7 @@ func (m multiOutputHandler) Enabled(ctx context.Context, level slog.Level) bool 
 	return false
 }
 
-func (m multiOutputHandler) Handle(ctx context.Context, record slog.Record) error {
+func (m Handler) Handle(ctx context.Context, record slog.Record) error {
 	var errs []error
 	for _, handler := range m {
 		if handler.Enabled(ctx, record.Level) {
@@ -27,18 +27,18 @@ func (m multiOutputHandler) Handle(ctx context.Context, record slog.Record) erro
 	return errors.Join(errs...)
 }
 
-func (m multiOutputHandler) WithAttrs(attrs []slog.Attr) slog.Handler {
+func (m Handler) WithAttrs(attrs []slog.Attr) slog.Handler {
 	var newHandlers []slog.Handler
 	for _, handler := range m {
 		newHandlers = append(newHandlers, handler.WithAttrs(attrs))
 	}
-	return multiOutputHandler(newHandlers)
+	return Handler(newHandlers)
 }
 
-func (m multiOutputHandler) WithGroup(name string) slog.Handler {
+func (m Handler) WithGroup(name string) slog.Handler {
 	var newHandlers []slog.Handler
 	for _, handler := range m {
 		newHandlers = append(newHandlers, handler.WithGroup(name))
 	}
-	return multiOutputHandler(newHandlers)
+	return Handler(newHandlers)
 }
